WIP: refactor, cleanup, and fixes related to most recent PR review notes
